refactor(host): split agent.rs into per-provider modules
Break agent.rs into agent_claude, agent_codex, agent_opencode for
provider-specific logic; shared helpers live in agent_utils. Rename
claude_cli_probe to agent_claude_probe and fold account_usage into
agent_utils. Gate hooks setup/detection behind debug_assertions so
release builds skip hook install and report hooks_installed = false.
