+++ crates/zedra-host/src/agent_installed.rs
